Add draw circuit
